namespace forum, update stack accordingly
