+++ app/assets/stylesheets/application.css
@@ -9,6 +9,6 @@
  * compiled file, but it's generally better to create a new file per style scope.
  *
  *= require_self
- *= require_tree 
+ *= require_tree
  *= require bootstrap.min
  */
