+++ app/assets/stylesheets/carousel.css.scss
@@ -75,7 +75,7 @@ $main-color: white;
     width: 40vw;
     background-color: rgba(0,0,0,0.35);
     min-height: 80vh;
-    
+
 	.ball {
 		text-align: center;
 		min-height: 33vh;
