+++ app/assets/stylesheets/channels.css.scss
@@ -1,4 +1,4 @@
-// Place all the styles related to the Channels controller here.
+// Place all the styles related to the Channel controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
