+++ app/assets/stylesheets/welcome.css.scss
@@ -27,7 +27,7 @@ overflow-x: hidden;
 
 .div1 {
 	height: 10vh;
-	background-color: rgba(0,0,100,1);  
+	background-color: rgba(0,0,100,1);
 	& .cal-text a {
 		color: white !important;
 	}
