--- app/controllers/messages_controller.rb
@@ -1,20 +1,0 @@
-class MessagesController < ApplicationController
-  
-  def index
-    @messages = Message.all
-  end
-  
-  def new
-    @message = Message.new
-  end
-  
-  def create
-    @message = Message.new(message_params)
-  end
-  
-  private
-  def message_params
-    params.require(:message).permit(:content)
-  end
-  
-end
