--- app/helpers/channels_helper.rb
@@ -1,2 +1,0 @@
-module ChannelsHelper
-end
