--- app/models/channels.rb
@@ -1,3 +1,0 @@
-class Channels < ActiveRecord::Base
-  has_many :messages
-end
