+++ app/models/forum/channel.rb
@@ -1,0 +1,3 @@
+class Forum::Channel < ActiveRecord::Base
+  has_many :messages, class_name: 'Forum::Message'
+end
