+++ app/models/forum/message.rb
@@ -1,0 +1,3 @@
+class Forum::Message < ActiveRecord::Base
+  belongs_to :channel, class_name: 'Forum::Channel'
+end
