--- app/models/message.rb
@@ -1,3 +1,0 @@
-class Message < ActiveRecord::Base
-  belongs_to :channel
-end
