+++ app/views/devise/registrations/new.html.erb
@@ -1,48 +1,47 @@
 <div class='jumbotron jt4'>
-	<div class='container-fluid signup'>
-		<h2 class='blurb2'>Sign up</h2>
+  <div class='container-fluid signup'>
+    <h2 class='blurb2'>Sign up</h2>
 
-		<%= form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
-		  <%= devise_error_messages! %>
-		  
-		  <div class="field">
-			<%= f.label :first_name %><br />
-			<%= f.text_field :first_name %>
-		  </div>
-	
-		  <div class="field">
-			<%= f.label :last_name %><br />
-			<%= f.text_field :last_name %>
-		  </div>
+    <%= form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
+      <%= devise_error_messages! %>
 
-		  <div class="field">
-		    <%= f.label :email %><br />
-		    <%= f.email_field :email, autofocus: true %>
-		  </div>
+      <div class="field">
+        <%= f.label :first_name %><br />
+        <%= f.text_field :first_name, autofocus: true %>
+      </div>
 
-		  <div class="field">
-		    <%= f.label :password %>
-		    <% if @minimum_password_length %>
-		    <em>(<%= @minimum_password_length %> characters minimum)</em>
-		    <% end %><br />
-		    <%= f.password_field :password, autocomplete: "off" %>
-		  </div>
+      <div class="field">
+        <%= f.label :last_name %><br />
+        <%= f.text_field :last_name %>
+      </div>
 
-		  <div class="field">
-		    <%= f.label :password_confirmation %><br />
-		    <%= f.password_field :password_confirmation, autocomplete: "off" %>
-		  </div>
+      <div class="field">
+        <%= f.label :email %><br />
+        <%= f.email_field :email%>
+      </div>
 
-		  <div class="actions blurb2">
-		    <%= f.submit "Sign up", :class => 'su-button' %>
-		  </div>
-		<% end %>
-		
-		<div class="blurb2">
-			<%= render "devise/shared/links" %>
-		</div>
-	</div>
+      <div class="field">
+        <%= f.label :password %>
+        <% if @minimum_password_length %>
+          <em>(<%= @minimum_password_length %> characters minimum)</em>
+        <% end %><br />
+        <%= f.password_field :password, autocomplete: "off" %>
+      </div>
+
+      <div class="field">
+        <%= f.label :password_confirmation %><br />
+        <%= f.password_field :password_confirmation, autocomplete: "off" %>
+      </div>
+
+      <div class="actions blurb2">
+        <%= f.submit "Sign up", :class => 'su-button' %>
+      </div>
+    <% end %>
+
+    <div class="blurb2">
+      <%= render "devise/shared/links" %>
+    </div>
+  </div>
 </div>
 
-<%= render file: '/layouts/balls' %>
-
+<%= render file: 'layouts/balls' %>
