+++ app/views/forum/channels/index.html.erb
