+++ app/views/forum/channels/new.html.erb
@@ -1,6 +1,6 @@
 <div class="container-fluid">
 	<%= form_for @channel do |f| %>
-		<%= f.text_area :content %>
-		<%= f.submit "Post"%>
+		<%= f.text_area :topic %>
+		<%= f.submit "Save"%>
 	<% end %>
 </div>
