+++ app/views/forum/channels/show.html.erb
