--- app/views/layouts/         Previous   .desktop
@@ -1,6 +1,0 @@
-[Desktop Entry]
-Encoding=UTF-8
-Name=Link to          Previous   
-Type=Link
-URL=http://0.0.0.0:3000/#myCarousel
-Icon=text-html
