+++ app/views/layouts/carousel.html.erb
@@ -13,7 +13,7 @@
 
   <!-- Wrapper for slides -->
   <div class="carousel-inner" role="listbox">
-  
+
     <div class="item item1 active hero-container">
         <div class="title">
             <div class="logo">
@@ -23,7 +23,7 @@
               <hr class="blurb line">
               <h3 class="blurb">Mitch Murphy Youth Sports Inc.</h3>
           </div>
-             
+
 
         </div>
     </div>
@@ -37,11 +37,11 @@
 		                	<title>baseball</title>
 		                	<path class="path1" d="M25 25.665c-2.23 2.069-5.217 3.335-8.5 3.335s-6.27-1.265-8.5-3.335v0 0c0.541-0.502 1.038-1.052 1.484-1.643l0.736 0.535 0.588-0.809-0.753-0.547c0.496-0.78 0.909-1.619 1.224-2.504l0.93 0.302 0.309-0.951-0.942-0.306c0.236-0.879 0.377-1.797 0.415-2.742h1.010v-1h-1.010c-0.037-0.945-0.179-1.862-0.415-2.742l0.942-0.306-0.309-0.951-0.93 0.302c-0.315-0.885-0.728-1.724-1.224-2.504l0.753-0.547-0.588-0.809-0.736 0.534c-0.446-0.591-0.943-1.14-1.484-1.643v0c2.23-2.069 5.217-3.335 8.5-3.335s6.27 1.265 8.5 3.335v0c-0.541 0.502-1.038 1.052-1.484 1.643l-0.736-0.534-0.588 0.809 0.753 0.547c-0.496 0.78-0.909 1.619-1.224 2.504l-0.93-0.302-0.309 0.951 0.942 0.306c-0.235 0.879-0.377 1.797-0.415 2.742h-1.010v1h1.010c0.037 0.945 0.179 1.862 0.415 2.742v0l-0.942 0.306 0.309 0.951 0.93-0.302c0.315 0.885 0.728 1.724 1.224 2.504l-0.753 0.547 0.588 0.809 0.736-0.535c0.446 0.591 0.942 1.14 1.484 1.643v0 0 0zM26.267 24.301c1.71-2.138 2.733-4.85 2.733-7.801 0-3.261-1.249-6.23-3.294-8.456-0.504 0.465-0.966 0.974-1.381 1.521l0.882 0.641-0.588 0.809-0.863-0.627c-0.437 0.695-0.801 1.44-1.083 2.224l0.971 0.316-0.309 0.951-0.959-0.312c-0.205 0.781-0.331 1.595-0.366 2.432h0.989v1h-0.989c0.036 0.837 0.161 1.651 0.366 2.432l0.959-0.311 0.309 0.951-0.971 0.316c0.282 0.784 0.646 1.529 1.083 2.224l0.863-0.627 0.588 0.809-0.882 0.641c0.414 0.547 0.877 1.057 1.381 1.521-0.014 0.015 0.369-0.414 0.562-0.655v0zM6.645 24.19c-1.657-2.121-2.645-4.79-2.645-7.69 0-3.261 1.249-6.23 3.294-8.456 0.504 0.465 0.966 0.974 1.381 1.521l-0.882 0.641 0.588 0.809 0.863-0.627c0.437 0.695 0.801 1.44 1.083 2.224l-0.971 0.316 0.309 0.951 0.959-0.312c0.205 0.781 0.331 1.595 0.366 2.432h-0.989v1h0.989c-0.036 0.837-0.161 1.651-0.366 2.432l-0.959-0.311-0.309 0.951 0.971 0.316c-0.282 0.784-0.646 1.529-1.083 2.224l-0.863-0.627-0.588 0.809 0.882 0.641c-0.414 0.547-0.877 1.057-1.381 1.521 0.014 0.015-0.431-0.486-0.649-0.765v0z">
 	                		</path>
-	              		</symbol>    
+	              		</symbol>
 	            	</use>
 	        	</svg>
 	        </div>
-	        
+	
 	        <div class="blurb">
     			<h2 class='sidetxt'>Guiding Today's Youth Through Competitve Sports In Safe Environments While Providing Quality Equipment</h2>
     		</div>
@@ -61,7 +61,7 @@
           		</use>
         	</svg>
         </div>
-        
+
         <div class="blurb">
         	<h2 class='sidetxt'>Promoting Opportunities in Today's Next-Level Elite Youth Sports</h2>
     	</div>
@@ -81,7 +81,7 @@
           		</use>
         	</svg>
     	</div>
-        
+
         <div class="blurb">
         	<h2 class='sidetxt'>Proud Supporters of Promoting Opportunity in Today's Next-Level Elite Youth Sports</h2>
     	</div>
