+++ app/views/layouts/navbar.html.erb
@@ -17,14 +17,14 @@
         <li class="dropdown nav-item">
           <a href="#" class="dropdown-toggle blurb" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Sign up<span class="caret"></span></a>
           <ul class="dropdown-menu">
-            
+
 	    <% if !user_signed_in? %>
 		<li><a class='blurb' href="/users/sign_up" title="Sign_up">Sign up</a></li>
             	<li><a class='blurb' href="/users/sign_in" title="Log_in">Log In</a></li>
 	    <% else %>
 		<li class='blurb'><%= link_to "Log Out", destroy_user_session_path, :method => :delete %></li>
 	    <% end %>
-		           
+		
             <!--<li><a class='blurb' href="/users/log_out" title="Log_out">Log out</a></li>-->
           </ul>
         </li>
