+++ config/routes.rb
@@ -7,18 +7,23 @@ MMYS::Application.routes.draw do
   get '/sponsors/' => 'sponsors#index'
   get '/contact' => 'contact#index'
   get '/sponsors/new' => 'sponsors#new'
-  
+
+  namespace :forum do
+    resources :channels, except: :update
+  end
+=begin
   get '/forum/channels' => 'channels#index'
-  get '/forum/channels/:id' => 'channels#show', as: :channel
   get '/forum/channels/new' => 'channels#new', as: :new_channel
-  
+  get '/forum/channels/:id' => 'channels#show', as: :channel
+=end
+
   # get '/signup' => 'users#new'
   # resources :users
 
   # get '/login' => 'sessions#new'
   # post 'login' => 'sessions#create'
   # delete 'logout' => 'sessions#destroy'
-  
+
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
