+++ db/migrate/20161016224933_create_sponsors.rb
@@ -4,7 +4,7 @@ class CreateSponsors < ActiveRecord::Migration
     	t.string :name
     	t.string :image
     	t.string :website
- 
+
 
       t.timestamps
     end
