+++ db/migrate/20170110030627_create_channels.rb
@@ -1,12 +1,12 @@
-class CreateChannels < ActiveRecord::Migration
+class CreateChannel < ActiveRecord::Migration
   def change
     create_table :channels do |t|
       t.text :topic
       t.integer :posts
 
       t.timestamps
-      
-      
+
+
     end
   end
 end
