+++ test/controllers/channels_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class ChannelsControllerTest < ActionController::TestCase
+class ChannelControllerTest < ActionController::TestCase
   # test "the truth" do
   #   assert true
   # end
